Morello: Use Morello registers in longjmp and setjmp
Preserve the full 129 bits of r17-r29 and CSP, as defined in Morello
extension to AAPCS64. sigsetjmp is modified to restore the full link
register after calling setjmp.

+++ include/setjmp.h
@@ -9,9 +9,11 @@ extern "C" {
 
 #include <bits/setjmp.h>
 
+#include <stdint.h>
+
 typedef struct __jmp_buf_tag {
 	__jmp_buf __jb;
-	unsigned long __fl;
+	uintptr_t __fl;
 	unsigned long __ss[128/sizeof(long)];
 } jmp_buf[1];
